# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample swagger for Fiber",
        "title": "mappa-api",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Guionardo Furlan",
            "email": "guionardo@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.4.0"
    },
    "basePath": "/",
    "paths": {
...
        "/tg/pub": {
            "post": {
                "description": "Publica mensagem em chat do Telegram",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "telegram"
                ],
                "summary": "Telegram Publisher handler",
                "parameters": [
                    {
                        "description": "Bot request data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.BotRequestData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ReplyMessage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ReplyMessage"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.BotRequestData": {
            "type": "object",
            "properties": {
                "cId": {
                    "description": "Chat ID",
                    "type": "integer"
                },
                "mId": {
                    "description": "Message ID to respond to",
                    "type": "integer"
                },
                "msg": {
                    "description": "Message",
                    "type": "string"
                }
            }
        },
        "handlers.ReplyMessage": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}